+++ src/App.js
@@ -2,17 +2,18 @@ import firebase from 'firebase/compat/app';
 import 'firebase/compat/auth';
 import 'firebase/compat/firestore';
 
+import { getDownloadURL, getStorage, ref, uploadBytes } from 'firebase/storage';
 import { useAuthState } from 'react-firebase-hooks/auth';
-import { useCollectionData } from 'react-firebase-hooks/firestore';
+import { v4 } from 'uuid';
 
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import './App.css';
 import './Root.css';
 import Authentication from './pages/Authentication';
 import Main from './pages/Main';
 import TopBar from './pages/TopBar';
 
-firebase.initializeApp({
+const app = firebase.initializeApp({
     apiKey: process.env.REACT_APP_API_KEY,
     authDomain: process.env.REACT_APP_AUTH_DOMAIN,
     projectId: process.env.REACT_APP_PROJECT_ID,
@@ -23,10 +24,11 @@ firebase.initializeApp({
 
 const auth = firebase.auth();
 const firestore = firebase.firestore();
+export const storage = getStorage(app);
 
 function App() {
     const [ user ] = useAuthState(auth)
-    const [sidebarToggle, setSidebarToggle] = useState(false);
+    
 
     const handleSignOut = async () => {
     try {
@@ -38,7 +40,10 @@ function App() {
     }
 
     const handleUpdateUserInfo = async (userInfo) => {
-        await auth.currentUser.updateProfile(userInfo)
+        auth.currentUser.updateProfile(userInfo).then(() => {
+            console.log('Updated user info successfully!')
+            window.location.reload()
+        })
         .catch((err) => console.error(err))
     }
 
@@ -61,6 +66,15 @@ function App() {
         </div>
     </div>
   );
+}
+
+function isValidUrl(url) {
+    try {
+        new URL(url);
+        return true;
+    } catch (error) {
+        return false;
+    }
 }
 
 export default App;